Close #9 : Tracking : Switch from measuring rectangle to use zoom level
Changed the GUI interface to have a new edition for 'scale threshold' :
'Threshold Scale for tracking [1:200]' (by default)
We can edit the scale with 2 rules:
(1) enter directly the scale 'int'
(2) enter scale in QGIS format '1:[int]' (we accept : [int]:[int])
After editing the value, we transform the edit text to:
'1:[int]' (if no error)
else we keep the old/default value for scale threshold.

TODO : need to use a RegExp for filtering the input user.