new branch (new strategy with Git Hub to organise our work) : develop-feature-optimize-threads
Goal : Optimize threads managing.

In this push (goal of this branch), we reduce drastically the numbers of calls to threads process. We used delays timers to optimize the 'waiting'/'stopping' delays for threads.
Calling a thread process is done only if needed (datas in queues/stacks).

This approach is still locks-free