resolve  #25
add a tool function extent_equal to imt_tools compute if two extents are the same with a epsilon distance value.
It seems to be difficult to compute exactly this epsilon to detect our zoom (for refreshmap). The epsilon is set to 0.01 meter. I suspect the zoom extent from QGIS to apply a zoom with a truncated zoom value.