Issue #45
New approach with 'Pickle' python module.
This module is designed for serialization in python langage.
The difficulty here comes from Qt, because Qt and class derived from Qt Objects (like our plugin) is a wrapped python implementation (from C++ API). It's pretty difficult to deal with wrapped objects like this.
In this commit, i introduce a first step to serialize GUI informations (checkbox states for example) into QSettings.
With this mechanism we can save/restore 'isChecked' state of checkbox (list of).
This tool can be useful (essential) to design a 'automatic' serialization of our plug (Qt-GUI and others attributes).